Refactor: move quiz models to quiz/models package
