+++ app/src/main/java/fr/free/nrw/commons/quiz/QuizActivity.java
@@ -14,7 +14,7 @@ import com.facebook.drawee.drawable.ProgressBarDrawable;
 import com.facebook.drawee.generic.GenericDraweeHierarchyBuilder;
 import com.facebook.drawee.view.SimpleDraweeView;
 
-import fr.free.nrw.commons.data.models.quiz.QuizQuestion;
+import fr.free.nrw.commons.quiz.models.QuizQuestion;
 import java.util.ArrayList;
 
 import butterknife.BindView;
