+++ app/src/test/kotlin/fr/free/nrw/commons/quiz/QuizQuestionTest.kt
@@ -2,7 +2,7 @@ package fr.free.nrw.commons.quiz
 
 import android.net.Uri
 import fr.free.nrw.commons.TestCommonsApplication
-import fr.free.nrw.commons.data.models.quiz.QuizQuestion
+import fr.free.nrw.commons.quiz.models.QuizQuestion
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
